fix: flyway migrations / Patient and Practitioner relations; add: Account, Appointment pkgs
