+++ backend/src/main/java/com/equipo2/healthtech/model/appointment/Appointment.java
@@ -1,30 +1,21 @@
-package com.equipo2.healthtech.model.user;
+package com.equipo2.healthtech.model.appointment;
 
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "practitioner_roles")
+@Table(name = "appointments")
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
 @EqualsAndHashCode(callSuper = false)
-public class PractitionerRole {
+public class Appointment {
 
     @Id()
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long practitionerId;
-
-    private Long roleCodeId;
-
-    @NotBlank
-    @Size(max = 255)
-    private String speciality;
 }
