+++ backend/src/main/java/com/equipo2/healthtech/model/appointment/AppointmentNotification.java
@@ -1,0 +1,21 @@
+package com.equipo2.healthtech.model.appointment;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "appointment_notifications")
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@EqualsAndHashCode(callSuper = false)
+public class AppointmentNotification {
+
+    @Id()
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+}
