+++ backend/src/main/java/com/equipo2/healthtech/model/encounter/Encounter.java
@@ -1,16 +1,16 @@
-package com.equipo2.healthtech.model.patientRecord;
+package com.equipo2.healthtech.model.encounter;
 
 import com.equipo2.healthtech.model.AuditableEntity;
 import jakarta.persistence.*;
 import lombok.*;
 
 @Entity
-@Table(name = "patientRecords")
+@Table(name = "encounters")
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
 @EqualsAndHashCode(callSuper = false)
-public class PatientRecord extends AuditableEntity {
+public class Encounter extends AuditableEntity {
 
     @Id()
     @GeneratedValue(strategy = GenerationType.IDENTITY)
