+++ backend/src/main/java/com/equipo2/healthtech/model/practitioner/PractitionerRole.java
@@ -1,0 +1,30 @@
+package com.equipo2.healthtech.model.practitioner;
+
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "practitioner_roles")
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@EqualsAndHashCode(callSuper = false)
+public class PractitionerRole {
+
+    @Id()
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private Long practitionerId;
+
+    private Long roleCodeId;
+
+    @NotBlank
+    @Size(max = 255)
+    private String speciality;
+}
