+++ backend/src/main/resources/db/migration/V1__users.sql
@@ -1,4 +1,4 @@
-CREATE TYPE role_enum AS ENUM ('ADMIN', 'SUPER_ADMIN', 'PATIENT', 'PROVIDER', 'STAFF', 'BILLING');
+CREATE TYPE role_enum AS ENUM ('ADMIN', 'SUPER_ADMIN', 'PATIENT', 'PRACTITIONER', 'STAFF');
 
 CREATE TABLE users (
     id BIGSERIAL PRIMARY KEY,
