+++ backend/src/main/resources/db/migration/V2__user_profiles.sql
@@ -2,6 +2,7 @@ CREATE TABLE user_profiles (
     id BIGSERIAL PRIMARY KEY,
     user_id BIGINT UNIQUE NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     full_name VARCHAR(255) NOT NULL,
+    gender VARCHAR(50) NOT NULL,
     phone VARCHAR(50) NOT NULL,
     address TEXT NOT NULL,
     birthday DATE NOT NULL
